move links outside of Home.tsx to Links.tsx, create transformed for customizable links, continue implementing SWR as global state manager, other misc changes
