+++ client/src/components/Links/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Links';
